requirements.txt
PyOpenGL>=3.1.0
numpy>=1.19.0
glfw>=2.0.0
pygame>=2.0.0
